feat: extract metadata helpers
